requirements.txt
fastapi==0.128.2
uvicorn==0.40.0
sqlmodel==0.0.32
fastapi-mail==1.6.2
python-dotenv==1.2.1
PyJWT==2.11.0
argon2-cffi==23.1.0
phonenumbers==9.0.24
razorpay==1.4.1